+++ src/modules/0.0.19/iasql_functions/index.ts
@@ -1,5 +1,5 @@
 /* THIS MODULE IS A SPECIAL SNOWFLAKE. DON'T LOOK AT IT FOR HOW TO WRITE A REAL MODULE */
-import { Context, ModuleBase } from '../../interfaces';
+import { ModuleBase } from '../../interfaces';
 import { IasqlOperationType } from './entity';
 
 class IasqlFunctions extends ModuleBase {
@@ -8,10 +8,5 @@ class IasqlFunctions extends ModuleBase {
     super.init();
   }
   iasqlOperationType = IasqlOperationType;
-  // TODO: here for testing purpose. To be delete it
-  // TODO: this also will be refactored using an rpc subtype in ModuleBase
-  async customCall(ctx: Context, arg1: string, arg2: string) {
-    return JSON.stringify([{ result: 'I have been called!', arg1, arg2, memo: ctx.memo }]);
-  }
 }
 export const iasqlFunctions = new IasqlFunctions();
